New Controllers: CetateniController, PubeleController
Refactored Logic
