+++ Proiect_final/Colectare_pubela/Colectare_pubela/Views/Home/Index.cshtml
@@ -20,15 +20,15 @@
                     </div>
                     <div class="card-body">
                         <div class="d-grid gap-3">
-                            <button onclick="location.href='@Url.Action("AdaugaCetatean", "Home")'" class="btn btn-primary btn-lg hover-dark shadow">
+                            <button onclick="location.href='@Url.Action("AdaugaCetatean", "Cetateni")'" class="btn btn-primary btn-lg hover-dark shadow">
                                 Add Citizen
                             </button>
 
-                            <button onclick="location.href='@Url.Action("AdaugaPubela", "Home")'" class="btn btn-secondary btn-lg hover-dark shadow">
+                            <button onclick="location.href='@Url.Action("AdaugaPubela", "Pubele")'" class="btn btn-secondary btn-lg hover-dark shadow">
                                 Add Dumpster
                             </button>
 
-                            <button onclick="location.href='@Url.Action("AtribuirePubela", "Home")'" class="btn btn-success btn-lg hover-dark shadow">
+                            <button onclick="location.href='@Url.Action("AtribuirePubela", "Pubele")'" class="btn btn-success btn-lg hover-dark shadow">
                                 Assign Dumpster
                             </button>
 
@@ -36,7 +36,7 @@
                                 Register Collection
                             </button>
 
-                            <button onclick="location.href='@Url.Action("ColectariCetatean", "Home")'" class="btn btn-info btn-lg hover-dark shadow">
+                            <button onclick="location.href='@Url.Action("ColectariCetatean", "Cetateni")'" class="btn btn-info btn-lg hover-dark shadow">
                                 See Citizen's Collections
                             </button>
                         </div>
